+++ lib/instabug_flutter.dart
@@ -19,3 +19,4 @@ export 'src/modules/surveys.dart';
 export 'src/utils/instabug_navigator_observer.dart';
 export 'src/utils/screen_loading/instabug_capture_screen_loading.dart';
 export 'src/utils/screen_loading/route_matcher.dart';
+export 'src/utils/screen_name_masker.dart' show ScreenNameMaskingCallback;